refactored, added @-links in config for slick imports
