+++ src/features/login/ui/auth-modal.tsx
@@ -1,6 +1,6 @@
 import { X } from 'lucide-react';
 import { useState } from 'react';
-import { supabase } from '../../lib/supabase';
+import { supabase } from '../../../lib/supabase';
 
 interface AuthModalProps {
   isOpen: boolean;
