+++ src/features/login/ui/sign-up.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { supabase } from '../../lib/supabase';
+import { supabase } from '../../../lib/supabase';
 
 export function SignUp() {
   const [email, setEmail] = useState('');
