+++ src/shared/weekplan-card.tsx
